+++ resources/views/components/button/burger-menu.blade.php
@@ -1,3 +1,0 @@
-<button x-data="{}" @click="$dispatch('open-sidebar')" {{ $attributes->merge(['class' => 'pxy-05 outline']) }}>
-    <x-gt-icon-menu class="icon" />
-</button>