+++ src/View/Components/Menu.php
@@ -20,12 +20,18 @@ class Menu extends Component
         public bool $withHeaders = false,       // display menu header
         public string $itemClass = '',
         public string $iconClass = '',
+        public bool $isNewMenuComponent = false,       // tmp variable
     ) {
         $this->file = getJsonFile(resource_path("navs/$this->filename.json"));
     }
 
     public function render()
     {
+        if ($this->isNewMenuComponent) {
+            // used for single menu
+            return view('gotime::components.menu-new')->with(['menu' => $this->getMenu($this->menuname)]);
+        }
+
         return $this->menuname
             ?  view('gotime::components.menu')->with(['menu' => $this->getMenu($this->menuname)])
             : view('gotime::components.menus')->with(['file' => $this->file]);
@@ -37,6 +43,17 @@ class Menu extends Component
     public function isActive($item)
     {
 
+        $x = '';
+        if (isset($item->url)) {
+            if (request()->is($item->url)) {
+                $x = 'danger';
+            } else {
+                $x = 'orange';
+            };
+        }
+
+        // dd($item);
+        // dd('here');
         if (isset($item->route_name)) {
             return request()->routeIs("$item->route_name*");
         } elseif (isset($item->url)) {
@@ -53,27 +70,20 @@ class Menu extends Component
     }
 
     /**
-     * Create url from item route name or url
+     * Generates relative url from menu item (object)
      */
-    public function getUrl($item): string
+    public function getUrl(object $item): string
     {
         if (!isset($item->route_name) && !isset($item->url)) {
             throw new Exception("There is no route name or url defined for the '$item->name' menu item");
         }
 
-        // if (empty($item->exclude_route)) {
-            if (isset($item->route_name)) {
+        if (isset($item->route_name)) {
+            return route($item->route_name, absolute: false);
+        } else {
+            return ltrim($item->url, '/'); // strip '/' to prevent errors
+        }
 
-                return route($item->route_name); // http://naykel.site/naykel/docs
-
-            } elseif (isset($item->url)) {
-
-                return url($item->url);     // http://naykel.site/about
-
-            } else {
-
-                throw new Exception($item->name . " menu item has a problem. It's likely missing the route in wep.php!");
-            }
-        // }
+        throw new Exception($item->name . " menu item has a problem. It's likely missing the route in wep.php!");
     }
 }